Move another 13 classes into specimen-main
