+++ specimen/src/org/labkey/specimen/actions/ManageRequestBean.java
@@ -7,10 +7,6 @@ import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
-import org.labkey.api.specimen.model.SpecimenRequestActor;
-import org.labkey.api.specimen.requirements.SpecimenRequest;
-import org.labkey.api.specimen.requirements.SpecimenRequestRequirement;
-import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
 import org.labkey.api.study.Location;
 import org.labkey.api.util.Button;
 import org.labkey.api.view.ActionURL;
@@ -20,7 +16,11 @@ import org.labkey.specimen.SpecimenRequestManager;
 import org.labkey.specimen.actions.SpecimenController.ImportVialIdsAction;
 import org.labkey.specimen.actions.SpecimenController.OverviewAction;
 import org.labkey.specimen.actions.SpecimenController.RemoveRequestSpecimensAction;
+import org.labkey.specimen.model.SpecimenRequestActor;
 import org.labkey.specimen.notifications.ActorNotificationRecipientSet;
+import org.labkey.specimen.requirements.SpecimenRequest;
+import org.labkey.specimen.requirements.SpecimenRequestRequirement;
+import org.labkey.specimen.requirements.SpecimenRequestRequirementProvider;
 import org.labkey.specimen.security.permissions.ManageRequestsPermission;
 
 import java.util.ArrayList;
