+++ specimen/src/org/labkey/specimen/actions/SpecimenController.java
@@ -64,12 +64,7 @@ import org.labkey.api.specimen.importer.SimpleSpecimenImporter;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.SpecimenComment;
-import org.labkey.api.specimen.model.SpecimenRequestActor;
 import org.labkey.api.specimen.query.SpecimenQueryView;
-import org.labkey.api.specimen.requirements.SpecimenRequest;
-import org.labkey.api.specimen.requirements.SpecimenRequestRequirement;
-import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
-import org.labkey.api.specimen.requirements.SpecimenRequestRequirementType;
 import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
 import org.labkey.api.specimen.security.permissions.ManageRequestSettingsPermission;
 import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
@@ -129,6 +124,7 @@ import org.labkey.specimen.RequestedSpecimens;
 import org.labkey.specimen.SpecimenManager;
 import org.labkey.specimen.SpecimenRequestManager;
 import org.labkey.specimen.model.ExtendedSpecimenRequestView;
+import org.labkey.specimen.model.SpecimenRequestActor;
 import org.labkey.specimen.model.SpecimenRequestEvent;
 import org.labkey.specimen.notifications.ActorNotificationRecipientSet;
 import org.labkey.specimen.notifications.DefaultRequestNotification;
@@ -137,6 +133,10 @@ import org.labkey.specimen.pipeline.SpecimenArchive;
 import org.labkey.specimen.pipeline.SpecimenBatch;
 import org.labkey.specimen.query.SpecimenEventQueryView;
 import org.labkey.specimen.query.SpecimenRequestQueryView;
+import org.labkey.specimen.requirements.SpecimenRequest;
+import org.labkey.specimen.requirements.SpecimenRequestRequirement;
+import org.labkey.specimen.requirements.SpecimenRequestRequirementProvider;
+import org.labkey.specimen.requirements.SpecimenRequestRequirementType;
 import org.labkey.specimen.security.permissions.ManageDisplaySettingsPermission;
 import org.labkey.specimen.security.permissions.ManageNewRequestFormPermission;
 import org.labkey.specimen.security.permissions.ManageNotificationsPermission;
