+++ specimen/src/org/labkey/specimen/importer/SpecimenSettingsImporter.java
@@ -24,11 +24,6 @@ import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.importer.RequestabilityManager;
 import org.labkey.api.specimen.location.LocationCache;
 import org.labkey.api.specimen.location.LocationImpl;
-import org.labkey.api.specimen.model.SpecimenRequestActor;
-import org.labkey.api.specimen.requirements.RequirementType;
-import org.labkey.api.specimen.requirements.SpecimenRequestRequirement;
-import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
-import org.labkey.api.specimen.requirements.SpecimenRequestRequirementType;
 import org.labkey.api.specimen.settings.DisplaySettings;
 import org.labkey.api.specimen.settings.RepositorySettings;
 import org.labkey.api.specimen.settings.RequestNotificationSettings;
@@ -42,6 +37,11 @@ import org.labkey.security.xml.GroupType;
 import org.labkey.specimen.SpecimenRequestManager;
 import org.labkey.specimen.SpecimenRequestManager.SpecimenRequestInput;
 import org.labkey.specimen.actions.ManageReqsBean;
+import org.labkey.specimen.model.SpecimenRequestActor;
+import org.labkey.specimen.requirements.RequirementType;
+import org.labkey.specimen.requirements.SpecimenRequestRequirement;
+import org.labkey.specimen.requirements.SpecimenRequestRequirementProvider;
+import org.labkey.specimen.requirements.SpecimenRequestRequirementType;
 import org.labkey.specimen.writer.SpecimenArchiveDataTypes;
 import org.labkey.study.xml.DefaultRequirementType;
 import org.labkey.study.xml.DefaultRequirementsType;
