+++ specimen/src/org/labkey/specimen/requirements/SpecimenRequestRequirement.java
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 
-package org.labkey.api.specimen.requirements;
+package org.labkey.specimen.requirements;
 
 import org.labkey.api.data.Container;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
-import org.labkey.api.specimen.model.SpecimenRequestActor;
 import org.labkey.api.study.StudyCachable;
 import org.labkey.api.util.MemTracker;
+import org.labkey.specimen.model.SpecimenRequestActor;
 
 /**
  * User: brittp
