+++ specimen/src/org/labkey/specimen/requirements/SpecimenRequestRequirementProvider.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.labkey.api.specimen.requirements;
+package org.labkey.specimen.requirements;
 
 import org.labkey.api.data.Container;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.specimen.SpecimenSchema;
-import org.labkey.api.specimen.model.SpecimenRequestActor;
+import org.labkey.specimen.model.SpecimenRequestActor;
 
 import java.util.Collection;
 import java.util.HashSet;
