+++ specimen/src/org/labkey/specimen/requirements/SpecimenRequestRequirementType.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.labkey.api.specimen.requirements;
+package org.labkey.specimen.requirements;
 
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.location.LocationImpl;
