+++ specimen/src/org/labkey/specimen/view/manageActorOrder.jsp
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 %>
-<%@ page import="org.labkey.api.specimen.model.SpecimenRequestActor"%>
-<%@ page import="org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider"%>
-<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageActorOrderAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageActorsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageActorOrderAction"%>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageActorsAction"%>
+<%@ page import="org.labkey.specimen.model.SpecimenRequestActor" %>
+<%@ page import="org.labkey.specimen.requirements.SpecimenRequestRequirementProvider" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
