+++ specimen/src/org/labkey/specimen/view/manageRequest.jsp
@@ -24,9 +24,6 @@
 <%@ page import="org.labkey.api.specimen.Vial" %>
 <%@ page import="org.labkey.api.specimen.location.LocationImpl"%>
 <%@ page import="org.labkey.api.specimen.location.LocationManager" %>
-<%@ page import="org.labkey.api.specimen.model.SpecimenRequestActor" %>
-<%@ page import="org.labkey.api.specimen.requirements.SpecimenRequestRequirement" %>
-<%@ page import="org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider" %>
 <%@ page import="org.labkey.api.specimen.settings.SettingsManager" %>
 <%@ page import="org.labkey.api.study.Location" %>
 <%@ page import="org.labkey.api.study.SpecimenService" %>
@@ -38,7 +35,6 @@
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.template.ClientDependencies" %>
 <%@ page import="org.labkey.specimen.SpecimenManager" %>
-<%@ page import="org.labkey.specimen.SpecimenModule" %>
 <%@ page import="org.labkey.specimen.SpecimenRequestManager" %>
 <%@ page import="org.labkey.specimen.actions.ManageRequestBean" %>
 <%@ page import="org.labkey.specimen.actions.ShowSearchAction" %>
@@ -53,6 +49,9 @@
 <%@ page import="org.labkey.specimen.actions.SpecimenController.ManageRequirementAction" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.RequestHistoryAction" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.SubmitRequestAction" %>
+<%@ page import="org.labkey.specimen.model.SpecimenRequestActor" %>
+<%@ page import="org.labkey.specimen.requirements.SpecimenRequestRequirement" %>
+<%@ page import="org.labkey.specimen.requirements.SpecimenRequestRequirementProvider" %>
 <%@ page import="java.util.Arrays" %>
 <%@ page import="java.util.List" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
