+++ study/api-src/org/labkey/api/specimen/SpecimenManager.java
@@ -40,7 +40,6 @@ import org.labkey.api.security.User;
 import org.labkey.api.specimen.location.LocationCache;
 import org.labkey.api.specimen.model.SpecimenComment;
 import org.labkey.api.specimen.model.SpecimenTablesProvider;
-import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
 import org.labkey.api.study.SpecimenService;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
@@ -277,7 +276,9 @@ public class SpecimenManager
         Table.delete(SpecimenSchema.get().getTableInfoSampleAvailabilityRule(), containerFilter);
         assert set.add(SpecimenSchema.get().getTableInfoSampleAvailabilityRule());
 
-        SpecimenRequestRequirementProvider.get().purgeContainer(c);
+        SpecimenMigrationService SMS = SpecimenMigrationService.get();
+        if (null != SMS)
+            SMS.purgeRequestRequirementsAndActors(c);
         assert set.add(SpecimenSchema.get().getTableInfoSampleRequestRequirement());
         assert set.add(SpecimenSchema.get().getTableInfoSampleRequestActor());
 
@@ -313,7 +314,6 @@ public class SpecimenManager
         // VIEW: if this view gets removed, remove this line
         assert set.add(SpecimenSchema.get().getSchema().getTable("LockedSpecimens"));
 
-        SpecimenMigrationService SMS = SpecimenMigrationService.get();
         if (null != SMS)
             SMS.clearGroupedValuesForColumn(c);
     }
